+++ app/Http/Controllers/GroupStudentController.php
@@ -13,6 +13,7 @@ use App\Models\StudentAchievement;
 use App\Services\GroupStudentService;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Support\Carbon;
+use Illuminate\Support\Facades\Gate;
 use PhpOffice\PhpWord\TemplateProcessor;
 use Inertia\Inertia;
 
@@ -20,11 +21,13 @@ class GroupStudentController extends Controller
 {
     public function index(Group $group)
     {
+        Gate::authorize('viewAny', Group::class);
+
         $group->append('name');
         $group->load(['students.address', 'students.studyAddress', 'students.expulsion']);
 
         $group->students->each(function (Student $student) {
-            $student->append('age');
+            $student->append('age', 'can');
             $student->address?->append('address');
             $student->studyAddress?->append('address');
         });
@@ -34,12 +37,14 @@ class GroupStudentController extends Controller
 
     public function create(Group $group)
     {
+        Gate::authorize('create', Student::class);
         $group->append('name');
         return Inertia::render('group-student/CreatePage', compact('group'));
     }
 
     public function store(Group $group, GroupStudentRequest $request, GroupStudentService $groupStudentService)
     {
+        Gate::authorize('update', $group);
         $groupStudentService->create($group, $request);
         return to_route('groups.students.index', ['group' => $group->id]);
     }
@@ -47,6 +52,7 @@ class GroupStudentController extends Controller
     public function show(Group $group, string $studentNumber)
     {
         $student = $group->findStudentByNumber($studentNumber);
+        Gate::authorize('view', $student);
 
         $student->load([
             'address',
@@ -85,6 +91,7 @@ class GroupStudentController extends Controller
 
     public function edit(Group $group, string $studentNumber)
     {
+        Gate::authorize('update', $group);
         $student = $group->findStudentByNumber($studentNumber);
         $student->load(['address', 'studyAddress', 'passport']);
         $group->append('name');
@@ -98,12 +105,14 @@ class GroupStudentController extends Controller
         GroupStudentRequest $request,
         GroupStudentService $groupStudentService
     ) {
+        Gate::authorize('update', $group);
         $groupStudentService->update($student, $request);
         return to_route('groups.students.index', ['group' => $group->id]);
     }
 
     public function destroy(Group $group, Student $student, GroupStudentService $groupStudentService)
     {
+        Gate::authorize('delete', $group);
         $groupStudentService->delete($student);
         return to_route('groups.students.index', ['group' => $group->id]);
     }
@@ -111,6 +120,7 @@ class GroupStudentController extends Controller
     public function print(Group $group, string $studentNumber)
     {
         $student = $group->findStudentByNumber($studentNumber);
+        Gate::authorize('view', $student);
         $templateProcessor = new TemplateProcessor(resource_path('documents/student-card.docx'));
 
         $templateProcessor->setValues([
